requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pyserini>=0.22.0
weaviate-client>=3.25.0
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.6.0
numpy>=1.24.4
pandas>=2.0.3
tqdm>=4.66.1
pydantic>=2.4.2